% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regresion_logistica.R
\name{plot.logistico_greenreg}
\alias{plot.logistico_greenreg}
\title{Generación de Gráficas para Regresión Logística}
\usage{
\method{plot}{logistico_greenreg}(x, ...)
}
\description{
Generación de Gráficas para Regresión Logística
}
