% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regresion_lineal_simple.R
\name{plot.rls_greenreg}
\alias{plot.rls_greenreg}
\title{Gráficas de diagnóstico para rls_greenreg}
\usage{
\method{plot}{rls_greenreg}(x, ...)
}
\arguments{
\item{x}{Objeto de clase rls_greenreg.}

\item{...}{Argumentos adicionales.}
}
\description{
Genera una secuencia de 8 gráficas para validar el modelo.
}
