% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelo_Autorregresivo_(AR).R
\name{print.ar_greenreg}
\alias{print.ar_greenreg}
\title{Impresión para Modelos Autorregresivos (AR)}
\usage{
\method{print}{ar_greenreg}(x, ...)
}
\description{
Impresión para Modelos Autorregresivos (AR)
}
