% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Modelo_Autoregresivo_Integrado_de_Media_Movil_(ARIMA).R
\name{print.arima_greenreg}
\alias{print.arima_greenreg}
\title{Impresión Académica para Modelos Integrados ARIMA(p, d, q)}
\usage{
\method{print}{arima_greenreg}(x, ...)
}
\description{
Impresión Académica para Modelos Integrados ARIMA(p, d, q)
}
