% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelo_Autoregresivo_de_media_movil_(ARMA).R
\name{print.arma_greenreg}
\alias{print.arma_greenreg}
\title{Impresión para Modelo ARMA}
\usage{
\method{print}{arma_greenreg}(x, ...)
}
\description{
Impresión para Modelo ARMA
}
