% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regresion_logistica.R
\name{print.logistico_greenreg}
\alias{print.logistico_greenreg}
\title{Impresión para Regresión Logística}
\usage{
\method{print}{logistico_greenreg}(x, ...)
}
\description{
Impresión para Regresión Logística
}
