% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelo_media_movil_(MA).R
\name{print.ma_greenreg}
\alias{print.ma_greenreg}
\title{Impresión Académica para Modelos de Media Móvil (MA)}
\usage{
\method{print}{ma_greenreg}(x, ...)
}
\description{
Impresión Académica para Modelos de Media Móvil (MA)
}
