% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regresion_lineal_multiple.R
\name{print.rlm_greenreg}
\alias{print.rlm_greenreg}
\title{Reporte de Regresión Lineal Múltiple}
\usage{
\method{print}{rlm_greenreg}(x, ...)
}
\arguments{
\item{x}{Objeto de clase rlm_greenreg.}

\item{...}{Argumentos adicionales.}
}
\description{
Imprime un resumen detallado del modelo RLM.
}
