% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regresion_lineal_simple.R
\name{print.rls_greenreg}
\alias{print.rls_greenreg}
\title{Impresión de resultados rls_greenreg}
\usage{
\method{print}{rls_greenreg}(x, ...)
}
\arguments{
\item{x}{Objeto de clase rls_greenreg.}

\item{...}{Argumentos adicionales.}
}
\description{
Método optimizado para imprimir el reporte de la regresión.
}
