% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regresión_poisson.R
\name{print.poisson_greenreg}
\alias{print.poisson_greenreg}
\title{Impresión Académica para Regresión de Poisson}
\usage{
\method{print}{poisson_greenreg}(x, ...)
}
\description{
Impresión Académica para Regresión de Poisson
}
